tests/test-api/pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Joseph Rana <josephranamgr7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-dotenv = "^0.5.2"
requests = "^2.32.3"
pydantic = "^2.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
